dlf_backend_full/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pydantic
python-jose[cryptography]
passlib[bcrypt]
faker
python-dotenv
